refactor: rename util file to AuthUtil
